feat(amqp): add support for the json content-type
If you just pass data as the body to rhea, it gets serialized with their types.js,
and will make it through to other rhea clients, but it's not readable by anything else.

BREAKING CHANGE: The default content-type for amqp messages is switched to application/json,
but can still work as it used to with AMQP_MESSAGE_TYPE=rhea.