fix(mariadb): fix database deadlocks on insert
The SELECT ... FOR UPDATE locks more than necessary when the record isn't found, which was creating
deadlocks even between inserts with different ids. By committing and releasing the lock when the
record isn't found, we will get duplicate entry errors, but only if it's the same id. Since we
already have retry logic in place for duplicates and deadlocks, the update still goes through.